+++ generic/generic_utils.h
@@ -17,4 +17,4 @@ void change_dir( char *file );
 void exit_error( int exit_code, const char *fmt, ... );
 int set_rand_seed(); 
 
-double bytes_human(int bytes, std::string &unit_str);
+double bytes_human(size_t bytes, std::string &unit_str);
